refactor(web): breakdown-Accordion-component
